# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: 厂智访客 API 契约
  description: >
    工厂访客与车辆管理多租户 SaaS。分层前缀：/api/ops（运营平台）、/api/tenant（租户端）、
    /api/tenant/steel（钢铁专属）、/api/mobile（移动/员工）、/api/public（免登录）、/api/device（设备回调）。
    统一响应体 R<T>，分页 PageResult<T>。此契约覆盖主要端点，实现时按 docs/01 补全子资源与字段。
  version: 1.0.0
servers:
  - url: https://api.faccess.cn
    description: 生产
  - url: http://localhost:8080
    description: 本地
...
security:
  - bearerAuth: []
paths:
...
  /api/tenant/invitations:
    get:  { tags: [tenant-invitation], summary: 邀请记录, responses: { '200': { $ref: '#/components/responses/Ok' } } }
    post: { tags: [tenant-invitation], summary: 发起邀请（生成邀请码/二维码并发送）, requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/InvitationReq' } } } }, responses: { '200': { $ref: '#/components/responses/Ok' } } }
  /api/tenant/invitations/{id}/resend: { post: { tags: [tenant-invitation], summary: 重发邀请, parameters: [ { $ref: '#/components/parameters/Id' } ], responses: { '200': { $ref: '#/components/responses/Ok' } } } }
  /api/tenant/invitations/{id}/cancel: { post: { tags: [tenant-invitation], summary: 取消邀请, parameters: [ { $ref: '#/components/parameters/Id' } ], responses: { '200': { $ref: '#/components/responses/Ok' } } } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    deviceAuth: { type: apiKey, in: header, name: X-Device-Sign, description: 设备ID+HMAC签名+时间戳+nonce }
  parameters:
    Id:   { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
...
  responses:
    Ok:
      description: OK
      content: { application/json: { schema: { $ref: '#/components/schemas/R' } } }
  schemas:
    R:
      type: object
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: success }
        data: {}
        traceId: { type: string }
...
    InvitationReq:
      type: object
      required: [visitorName, visitorMobile, reason]
      properties:
        visitorName: { type: string }
        visitorMobile: { type: string }
        company: { type: string }
        reason: { type: string }
        hostUserId: { type: integer, format: int64 }
        companions: { type: integer }
        visitDate: { type: string, format: date }
        timeFrom: { type: string }
        timeTo: { type: string }
        sendChannels: { type: array, items: { type: string, enum: [SMS, WECHAT, LINK] } }
...